add basic getcpucount & delint
git-svn-id: https://svn.aros.org/svn/aros/trunk@50373 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ arch/arm-native/kernel/kernel_arm.h
@@ -13,7 +13,7 @@ struct ARM_Implementation
     void                (*ARMI_Delay) (int);
     unsigned int        (*ARMI_GetTime) (void);
     void                (*ARMI_PutChar) (int);
-    void                (*ARMI_SerPutChar) (int);
+    void                (*ARMI_SerPutChar) (uint8_t);
     int                 (*ARMI_SerGetChar) (void);
     void                (*ARMI_IRQInit) ();
     void                (*ARMI_IRQEnable) (int);
